+++ src/Pages/MyPost/WritePost/WritePost.js
@@ -1,22 +1,116 @@
-import { React, useState } from "react";
+import { React, useState, useLayoutEffect } from "react";
 import styled from "styled-components";
 import "./WritePost.css";
 
 import theme from "../../../Styles/theme";
-import { FullSizeBtn } from "../../../Components/etc/Buttons";
+
 import BookInfo from "../../../Components/layout/BookInfo";
+import { bookSearch } from "APIs/api";
+import Item from "Components/layout/ListItem";
+import ModalFrame from "Components/layout/ModalFrame";
+import { FullSizeBtn, SmallBtn } from "Components/etc/Buttons";
+import { SearchBar } from "Components/etc/SearchBar";
+
+import { useDispatch, useSelector } from "react-redux";
+import { setQuery, setBooks, setPage, nextPage, isEndPage } from "Store/store";
 
 import { CKEditor } from "@ckeditor/ckeditor5-react";
 import ClassicEditor from "@ckeditor/ckeditor5-build-classic";
 import ReactHtmlParser from "react-html-parser";
 
 const WritePost = () => {
-  //글 내용 state들
+  //저장소에서 책검색 데이터 읽어오기
+  const queryData = useSelector((state) => state.bookSearch.query);
+  const bookList = useSelector((state) => state.bookSearch.books);
+  const pageNum = useSelector((state) => state.bookSearch.page);
+  const isEnd = useSelector((state) => state.bookSearch.isEnd);
+  const dispatch = useDispatch(); //작업 전달하기
+
+  ///////////////////////////////////////////  state 선언
+  //(임시) 글 내용 state들
   const [postContent, setPostContent] = useState({
     title: "",
     content: "",
   });
   const [viewContent, setViewContent] = useState([]);
+  //도서 state
+  const [modalState, setModalState] = useState(false); //모달
+  const [searchItem, setSearch] = useState("");
+  ///////////////////////////////////////////  state 선언 닫음
+
+  useLayoutEffect(() => {
+    //componentDidMount/Update/WillUnmount 일 경우 실행
+    //(query state가 업데이트되면 api 호출)
+    if (searchItem.length > 1) {
+      bookSearchHandler(queryData, pageNum);
+    }
+
+    if (!modalState) {
+      //초기화 처리
+      dispatch(setQuery(""));
+      dispatch(setPage(1));
+      dispatch(setBooks([]));
+      dispatch(isEndPage(true));
+    }
+    // eslint-disable-next-line react-hooks/exhaustive-deps
+  }, [queryData, pageNum]);
+
+  /////////////////////////////////책 검색용 함수들
+  const onClickSearch = () => {
+    dispatch(setQuery(searchItem));
+    dispatch(setPage(1));
+    dispatch(setBooks([]));
+  };
+
+  //엔터를 눌렀을 때 쿼리를 검색어로 교체하는 함수
+  const onEnter = (e) => {
+    if (e.keyCode === 13) {
+      onClickSearch();
+    }
+  };
+
+  //text 검색어가 바뀔 때 호출되는 함수.
+  const onTextUpdate = (e) => {
+    setSearch(e.target.value);
+  };
+
+  //책 검색
+  const bookSearchHandler = async (query, page) => {
+    const params = {
+      query: query, //검색어
+      sort: "accuracy", //accuracy: 정확도, latest: 발간일 순
+      page: page,
+      size: 10, //1~50. 출력할 검색 결과 수
+    };
+
+    const { data } = await bookSearch(params); //책 검색 api 호출
+    if (page === 1) {
+      dispatch(setBooks(data.documents));
+    } else if (page >= 2) {
+      dispatch(setBooks(bookList.concat(data.documents)));
+    }
+
+    dispatch(isEndPage(data.meta.is_end)); //다음 페이지가 있으면 false
+  };
+  /////////////////////////////////책 검색용 함수들 닫음
+
+  /////////////////////////////////모달용 함수들
+  const openModal = (e) => {
+    e.preventDefault();
+    setModalState(true);
+  };
+  const closeModal = (e) => {
+    e.preventDefault();
+    setModalState(false);
+
+    //모달창 닫기와 동시에 쿼리 초기화
+    setSearch("");
+    dispatch(setQuery(""));
+    dispatch(setPage(1));
+    dispatch(setBooks([]));
+    dispatch(isEndPage(true));
+  };
+  /////////////////////////////////모달용 함수들 닫음
 
   const getValue = (e) => {
     const { name, value } = e.target;
@@ -36,9 +130,8 @@ const WritePost = () => {
       ))}
 
       <div>
-        {/* <BookInfo /> */}
         <SubjectDiv>
-          <SubjectButton>
+          <SubjectButton onClick={openModal}>
             <ImgIcon src={require("Assets/icn_book.png")} alt="book" /> 도서
           </SubjectButton>
           <SubjectButton>
@@ -58,6 +151,8 @@ const WritePost = () => {
             onChange={getValue}
             name="title"
           />
+
+          {queryData && <BookInfo />}
 
           <CKEditor
             editor={ClassicEditor}
@@ -92,6 +187,41 @@ const WritePost = () => {
           </FullSizeBtn>
         </div>
       </div>
+
+      {modalState && (
+        <ModalFrame state={modalState} closeModal={closeModal}>
+          <SearchBar
+            value={searchItem}
+            onKeyDown={onEnter}
+            onChange={onTextUpdate}
+            onClick={onClickSearch}
+          />
+          <Blank />
+          {bookList.map((book, idx) => (
+            <Item
+              key={idx}
+              thumbnail={book.thumbnail}
+              title={book.title}
+              authors={book.authors}
+              datetime={book.datetime.substr(0, 4)}
+              publisher={book.publisher}
+              contents={book.contents}
+              tolink={"/my_post/write_post"}
+            />
+          ))}
+          <Blank />
+
+          {!isEnd && (
+            <SmallBtn
+              onClick={() => {
+                dispatch(nextPage());
+              }}
+            >
+              더보기
+            </SmallBtn>
+          )}
+        </ModalFrame>
+      )}
     </div>
   );
 };
@@ -127,6 +257,10 @@ const ImgIcon = styled.img`
   height: 35px;
   width: auto;
   margin-right: 8%;
+`;
+
+const Blank = styled.div`
+  height: 10px;
 `;
 
 export default WritePost;
